Added login apis

+++ user/src/main/java/com/user/service/Impl/UserServiceImpl.java
@@ -1,5 +1,7 @@
 package com.user.service.Impl;
 
+import com.commons.user.LoginDto;
+import com.commons.user.LoginResponseDto;
 import com.commons.user.UserDto;
 import com.commons.utility.PageableResponse;
 import com.commons.utility.SearchRequest;
@@ -9,7 +11,7 @@ import com.user.repo.UserRepository;
 import com.user.service.UserService;
 import com.user.service.UserUtilsService;
 import com.user.validator.UserValidator;
-import io.micrometer.common.util.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.util.Strings;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -56,12 +58,14 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserDto getById(String userId) {
-        return userMapper.userToUserDto(userRepository.findByUserId(userId));
+        if(!StringUtils.isBlank(userId))
+             return userMapper.userToUserDto(userRepository.findByUserId(userId));
+        return null;
     }
 
     @Override
     public boolean deleteById(String userId) {
-        if(userRepository.existsById(userId)){
+        if(!StringUtils.isBlank(userId) && userRepository.existsById(userId)){
             userRepository.deleteById(userId);
             return true;
         }
@@ -87,5 +91,30 @@ public class UserServiceImpl implements UserService {
         pageableResponse.setTotalRecord(allUsers.getTotalElements());
         pageableResponse.setData(userMapper.userListToUserDtoList(allUsers.getContent()));
         return pageableResponse;
+    }
+
+    @Override
+    public LoginResponseDto login(LoginDto loginDto) {
+        if(loginDto != null){
+            if(!StringUtils.isAnyBlank(loginDto.getPassword(),loginDto.getUserName())) {
+                User existingUser = userRepository.findByUsernameAndPassword(loginDto.getUserName(), loginDto.getPassword());
+                if (existingUser != null) {
+                    return setLoginResponse(existingUser.getFirstName() + " "+ existingUser.getLastName(), 200, existingUser.getRole(),"User logged in successfully.");
+                } else {
+                    return setLoginResponse(null, 500, null,"User does not exists for given username and password.");
+                }
+            }
+            return setLoginResponse(null,500,null,"Username or password cannot be empty.");
+        }
+        return null;
+    }
+
+    private LoginResponseDto setLoginResponse(String displayName ,Integer status,String role,String message) {
+        LoginResponseDto loginResponseDto = new LoginResponseDto();
+        loginResponseDto.setDisplayName(displayName);
+        loginResponseDto.setStatus(Integer.toString(status));
+        loginResponseDto.setRole(role);
+        loginResponseDto.setMessage(message);
+        return loginResponseDto;
     }
 }
